Major refactoring for better readability. No changes in functionality
