+++ .gitignore
@@ -3,3 +3,5 @@
 /BAD_segmentation.egg-info/
 /src/babachi.egg-info/
 /build
+**/__pycache__/
+*.bed
