+++ src/babachi/__main__.py
@@ -1,0 +1,4 @@
+from babachi.bad_estimation import segmentation_start
+
+if __name__ == '__main__':
+    segmentation_start()
