+++ src/babachi/logging.py
@@ -1,0 +1,15 @@
+import sys
+import logging
+
+
+def set_logger_config(logger, level):
+    logger.setLevel(level)
+    if not logger.handlers:
+        handler = logging.StreamHandler(sys.stdout)
+        handler.setLevel(level)
+        formatter = logging.Formatter('%(asctime)s  %(levelname)s  %(message)s')
+        handler.setFormatter(formatter)
+        logger.addHandler(handler)
+
+
+root_logger = logging.getLogger("babachi")
