+++ src/babachi/version.py
@@ -1,1 +1,1 @@
-__version__ = '2.0.26'
+__version__ = '3.0.0'
